Use quantity as 1 (or 0 for out-of-stock) for generic events
Events that are not cart/checkout/purhcase related, such as view/select item(s), do not need specify their `item`s `quantity` as other values than 1 or 0.

We'll use 1 to indicate the product is in-stock and 0 for out-of-stock.

Cart/checkout/purchase events items will still respect the user's selected quantity.