Take into account the item's quantity when calculating the value (#127)
#### What is the purpose of this pull request?

We were calculating the `value` of some events as if the `item`s `quantity` were always 1. I updated the mock data of the tests to use `quantity` other than 1 and assert `value` is taking the variable quantity into account. The fix is in [the last commit](https://github.com/vtex-apps/google-tag-manager/pull/127/commits/82c5f7256b9a3284f44a2af816affa6758b76529).

Affected events:
- `add_to_cart`, `remove_from_cart`, and `view_cart`
- `begin_checkout`
- `purchase`
- `refund`

We're calculating `value` for these events as the sum of the items' (sale) prices * quantity, e.g., if `view_cart` was triggered with a 2x $100 item and a 3x $400 item, `value` should be 1400 (`2 * 100 + 3 * 400`).