Fix many hidden type errors (#137)
#### What is the purpose of this pull request?

I noticed the `CartData` type inside the interface `PixelMessage` wasn't
defined anywhere in the project and decided to remove it.

<img width="479" alt="CleanShot 2023-04-25 at 16 32 45@2x"
src="https://user-images.githubusercontent.com/381395/234383504-ad76b220-37e8-4681-b85a-cb5a3c01652a.png">

Because it was unresolved, it took the type of `any`. This caused our
event types to lose effectiveness as `any` was too wide. It surfaced as
many as 31 type errors that we weren't handling! This PR addresses these
errors.

##### History on `CartData`

Originally added
[here](https://github.com/vtex-apps/google-tag-manager/pull/38/files#diff-791619406f30fb7292cdecb798d2a840d54a926399fdcd740a7706dbe852c9a6),
it was later
[renamed](https://github.com/vtex-apps/google-tag-manager/pull/40/files#diff-791619406f30fb7292cdecb798d2a840d54a926399fdcd740a7706dbe852c9a6)
to `CartLoadedData` but `CartData`
[remained](https://github.com/vtex-apps/google-tag-manager/blob/f785e969c4b947a8c699598c46bdefde3cca79a2/react/typings/events.d.ts#L17)
in the `PixelMessage` list, when it should have been removed.

#### Screenshots or example usage

Before|After
-|-

![tsc-before2](https://user-images.githubusercontent.com/381395/234426791-cefd1007-3c0f-406e-92e5-0df510e5c67d.png)|![tsc-after](https://user-images.githubusercontent.com/381395/234418290-ec627ce7-7054-4aa0-ac1b-26feab9906ab.png)

#### Types of changes

* [x] Bug fix (a non-breaking change which fixes an issue)
* [ ] New feature (a non-breaking change which adds functionality)
* [ ] Breaking change (fix or feature that would cause existing
functionality to change)
* [ ] Requires change to documentation, which has been updated
accordingly.